+++ common/src/main/java/com/alien/common/gameplay/entity/living/alien/ovomorph/ai/OvomorphGOAP.java
@@ -1,0 +1,35 @@
+package com.alien.common.gameplay.entity.living.alien.ovomorph.ai;
+
+import com.alien.common.gameplay.entity.living.alien.ovomorph.Ovomorph;
+import com.alien.common.model.alien.HatchState;
+import com.lib.common.gameplay.entity.ai.EntityGOAP;
+import com.lib.common.gameplay.goap.TypedIdentifier;
+import com.lib.common.gameplay.goap.state.GOAPMutableWorldState;
+
+public class OvomorphGOAP extends EntityGOAP<Ovomorph> {
+
+    public static final TypedIdentifier<HatchState> HATCH_STATE = new TypedIdentifier<>("hatchState");
+
+    public static final TypedIdentifier<Boolean> WANTS_TO_HATCH = new TypedIdentifier<>("wantsToHatch");
+
+    public OvomorphGOAP(Ovomorph ovomorph) {
+        addBaseRoutines(ovomorph);
+    }
+
+    public void addBaseRoutines(Ovomorph ovomorph) {
+        // Track the ovomorph's desire to hatch.
+        addSensor(new WantsToHatchSensor());
+        // Track the ovomorph's hatch state.
+        addSensor(this::senseHatchState);
+
+        // The desired goal of all ovomorphs is to hatch.
+        addGoal(new HatchGoal());
+
+        // How can the ovomorph hatch?
+        addAction(new HatchAction());
+    }
+
+    private void senseHatchState(Ovomorph ovomorph, GOAPMutableWorldState worldState) {
+        worldState.set(OvomorphGOAP.HATCH_STATE, ovomorph.getHatchState().unwrapOr(Ovomorph.DEFAULT_HATCH_STATE));
+    }
+}